Fixed routes and organized folders
