+++ src/App.jsx
@@ -1,18 +1,22 @@
-import './App.css'
-import Signup from './components/Signup'
-import MainPage from './components/MainPage.jsx'
-import Login from './components/Login.jsx'
+import './index.css'
+import Signup from './views/Signup'
+import MainPage from './views/MainPage.jsx'
+import Login from './views/Login.jsx'
 import { Link, Route, Routes } from 'react-router-dom'
-import Who from './components/who/Who'
+import Who from './views/Who'
 
 function App () {
   return (
     <div className='App'>
       <li><Link to='/'>MainPage</Link></li>
       <li><Link to='/Login'>Login</Link></li>
+      <li><Link to='/Signup'>Signup</Link></li>
+      <li><Link to='/Who'>Who</Link></li>
       <Routes>
         <Route path='/' element={<MainPage />}/>
         <Route path='/Login' element={<Login />}/>
+        <Route path='/Signup' element={<Signup />}/>
+        <Route path='/Who' element={<Who />}/>
       </Routes>
     </div>
   )
