+++ src/components/Button.jsx
@@ -1,4 +1,4 @@
-import './MainPage.css'
+import '../css/MainPage.css'
 import { Link } from 'react-router-dom'
 
 function Button (props) {
