--- src/components/MainPage.jsx
@@ -1,16 +1,0 @@
-import './MainPage.css'
-import Description from './Description.jsx'
-import Footer from './Footer.jsx'
-import LandingImage from './LandingImage'
-
-function MainPage () {
-  return (
-  <div>
-    <LandingImage />
-    <Description />
-    <Footer />
-  </div>
-  )
-}
-
-export default MainPage
