+++ src/views/MainPage.jsx
@@ -1,0 +1,16 @@
+import '../css/MainPage.css'
+import Description from '../components/Description'
+import Footer from '../components/Footer.jsx'
+import LandingImage from '../components/LandingImage'
+
+function MainPage () {
+  return (
+  <div>
+    <LandingImage />
+    <Description />
+    <Footer />
+  </div>
+  )
+}
+
+export default MainPage
